+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.2.0](https://github.com/ErezNagar/lease-calculator/compare/v2.1.0...v2.2.0) - 04/6/2021
+
+### Added
+
+- An ability to get a detailed list of all drive-off payments
+  - `getDriveOffPaymentDetails(): Array<Object> | null`
+
 ## [2.1.0](https://github.com/ErezNagar/lease-calculator/compare/v2.0.0...v2.1.0) - 31/1/2021
 
 ### Added